Chore: 주석 추가 및 전체적으로 코드 정리
