+++ src/main/java/umc/duckmelang/domain/auth/controller/AuthController.java
@@ -10,15 +10,12 @@ import umc.duckmelang.domain.auth.service.AuthService;
 import umc.duckmelang.global.apipayload.ApiResponse;
 import umc.duckmelang.global.redis.blacklist.BlacklistService;
 import umc.duckmelang.global.security.filter.JwtAuthorizationFilter;
-import umc.duckmelang.global.security.jwt.JwtTokenProvider;
 import umc.duckmelang.global.security.jwt.JwtUtil;
 
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
     private final AuthService authService;
-    private final JwtAuthorizationFilter jwtAuthorizationFilter;
-    private final JwtTokenProvider jwtTokenProvider;
     private final JwtUtil jwtUtil;
     private final BlacklistService blacklistService;
 
@@ -37,7 +34,7 @@ public class AuthController {
     @PostMapping("/logout")
     @Operation(summary = "로그아웃 API", description = "토큰을 블랙리스트에 추가합니다.")
     public ApiResponse<String> logout(HttpServletRequest request) {
-        String token = jwtAuthorizationFilter.extractToken(request);
+        String token = jwtUtil.extractToken(request);
         long expiration = jwtUtil.getExpirationFromToken(token);
         blacklistService.addToBlacklist(token, expiration);
         return ApiResponse.onSuccess("로그아웃되었습니다.");
